requirements.txt
selenium
click
pyyaml
PyPyDispatcher
sentry-sdk